Move InitializeBuilder from BuildManager into an extension
BuildManager creates NuGet manifests from project.json, but so will the new publish bundle commands. I needed somewhere to put this code so it could be used elsewhere. Extension method seemed tidy enough but open to suggestions.
